Initial Commit
I've got a bunch of doohickies running around, and a basic job-board
that matches requests and drives bots.

I took a short-cut where I made bots into nodes just so it's easier to
test, but my plan is to replace them with a particle shader later, so
they're built to function as data-only classes.