New Job Market
The old one had pressure at the sender and receiver side. A mine wanted
to get rid of stuff, and that pressure grew over time, because I didn't
want starvation among mines. And a factory wanted to receive stuff, and
that pressure grew over time to prevent starvation there.

But then I started thinking about the interface I wanted the actual
player to have to this whole system. I realized I wanted them to have
high-level goal-setting ability.

So they want to store X amount of stuff, and they want to have this many
bots, etc.
So what I realized is that starvation of mines is actually not really a
problem! Or at least not a problem for me, but instead for the player.

If there's no pressure to have this thing for any reason, then I should
mine it! Leave it where it is! And if you can satisfy all the
requirements of the system with only 2 mines and the other 10 aren't
required, then why would I send my bots to further away mines if I can
get by with these two closer ones?

So this is where I'm at right now. I think I know how to make it work
for warehouses and intermediate factories, but I haven't built that yet.