New Market Warehouses
Ok, Warehouses are now on the new system.
It exposed a flaw I hadn't noticed from my first implementation, which
was that I was using the "market price" for my Mine when making offers.
When all of my destinations were robot factories this was fine, because
they all had the same price, and that price was the market price, and it
worked out.

But once I started having warehouses with different weights it started
being really weird!
It's better now.

Also, for now I'm embracing the copy-and-paste for offers and requests
but I'm already starting to see some of the duplication, so I think I
know where I'll trim the fat later.

Cool!
It's hard to verify on an individual level what's going on in the
system, but I played around a bit and it didn't seem to do anything too
obviously stupid, or respond to stimuli in an obviously weird way, so
it's probably good enough there...