Added Second Commodity
Ok, the economy is vetted for now!
I have robots now requiring a secondary commodity (1) which can't be
mined, but can be produced in a red factory.

So now it moves things from the mine to the red factory, fills the
warehouses while that factory is working, and then when the factory is
done it moves things to the bot factory which makes bots!

Things don't make reds while the bot factory is working, because no
one's asking for reds.
And while the factories are working it loads up the warehouses. Then
when the factories come back on multiple bots all pull from the
warehouses at the same time to speed up the acquisition, and then they
refill the warehouses while the factories work!

Magic.