Bot Shader
I'm trying a thing where I use a shader to draw the bots floating around
and stuff. So now when the bot moves it updates the shader, but the
actual bot is drawn by the shader and a weird particle system.

I'm not fully there yet, so I've still got the shitty bot animating so I
can tell what's going on there. But the plan is to soon remove it so the
bot isn't even in the scene tree and is only drawn by the shader.