Acceleration Based Bot Positions
I now no-longer require the bots to be in the tree, and in fact have
turned it off by default (though it can be handy to turn on for
debugging so I've made it a flag)

A part of the benefit of that is that I also do a lot less work per-bot
in the CPU. I now basically just move them square to square along their
path, and the shader is the part that draws them smoothly.

That showed how they're actually not that smooth, though...
So I had to change it so they move with acceleration rather than just
velocity. Otherwise they can jerk pretty strangely when the position
changes the way it does, whereas now the velocity shift is smoother.

As a result of all of these changes the robots are much weirder. They
kinda bumble-about drunkenly, but for now I kinda like it!