Bot Shape and Colour
The bots are now glowy balls of colour.
This mostly works, though the blue looks more like purple to me for a
reason I don't need to understand right now.

In order to do this I had to do some kinda weird stuff to pass inventory
through the particle shader to the fragment shader, but it mostly worked
out.

When I did this I noticed how weird it was that the inventory became
full as soon as the virtual "bot" moved there, especially because the
rendered bot wouldn't have gotten there yet since it only just started
seeing that location as an option.

So I moved the logic around so that I move to a location, but I don't
process the pickup or dropoff until I've sat on it. Essentially I pickup
and dropoff when I leave the tile now, rather than when I enter it.

This seems to feel much more natural.