Fix Hover Reliability
I've been putting off looking into this thing to focus on the bots
stuff, knowing I'd be rebuilding mouse interactions anyway.

So it was super annoying, but not the most important thing.
But! It when I did look into it, it turns out it was actually really
simple! I just didn't initialize it so I had to enter, then leave, to
end up in the proper state. Then when I entereed again it would work.

But I hadn't noticed that pattern, because I wasn't looking into it. I
just felt like sometimes I'd have to leave a hex and go back before it
would work...