New Construction UI
I'm working towards building real construction, and this was the first
thing that had to change.

I now select objects, one or multiple, and then that gives me a UI to
click to choose what to build there.
No more cycling through things!

This is still basic, so the selecting mechanics aren't super robust and
there's no way to destroy something yet, but that's where we're at.