Construction
This isn't the way I want it to go in general, but it's a good first
step.

Now when I build the thing it first sits idle under the construction
cage, and then later the cage goes away and the building becomes active.

This involved separating adding the building to the tile from enabling
it to actually participate in the markets, etc.

In the future my plan is to have the construction actually cost
resources that need to be delivered rather than just 1 second, but this
is good enough for now to show that things can be in two states, etc.

The only thing to do now is add it to the market and I'll be there!