Refactor Market 1: Offers
I copy-and-pasted a bunch of code just to make sure things were working
and had properties I liked, and they did, but it lead to a bunch of
obvious duplication in the way offers worked.

But that was helpful in figuring out which things were common and which
things were special. Now, though, I think I've got a pretty ok handle on
it, so I'm cleaning this up.

I'm moving this more into a Node direction, as well. I think this
abstraction will help as time goes on to make things more modular. It
also takes a bunch of state which previously lived in variables and crap
and turning them more into implicit state. Like, when the timer fires I
create the offer, and when the offer is complete it deletes itself. That
means the offer can keep track of whether or not it's accepted, so it
can enter itself into the market or not. The parent doesn't have to
care. It just cares when the thing is gone.

So in that sense, the Offer node now functions as a kind of Inventory.
It's a thing that this node has that it's looking to get rid of, and
until that happens I just have it. Nice!

There's a few things in here that are not ideal, because Warehouses
still have their own Offer class. I'm going to get rid of those, but I
need some more pieces to do that which I'm planning on building next.

Once that's done, all Offers will have the consistent interface of the
one class, so I can start making assumptions based on that, etc.

While I was in here I noticed a bunch of other files that were hanging
around from my first attempt but weren't even being used anymore, so
I've deleted those.

Ok! So, next is requests!