Refactor Market 2: Requests, Slots
I've now built a more generic Request object rather than my more custom
version from before. This required some changes to my attempt from
before:

First, PriceSheet is out. Turns out as a resource it was a singleton, which
isn't really what I was looking for. Neat, but not what I wanted. So
I've replaced it with just a single dictionary, which is passed by
reference anyway, which honestly works better...

The second thing is that I previously relied on an intentional careful
sequencing to make sure my dependent pricing worked out. I would collect
all the requests raw, and then sort them after they were all in. That
way the pricing of one resource could be used to calculate another
assuming there wasn't any loops or anything, based on demand.

But that doesn't really work anymore here because I wanted things to be
more static. So instead I put a new specific signal into the Market loop
after all the requests are in and before offers are generated to
calculate prices.

That means the requests can go in, then the pricesheets can get updated
to reflect the current market price, and then the offers can be computed
based on those prices. Basically the same, but with better separation I
hope.

Then I made it so my custom Nodes (Offer, Request, Slot) all declare a
name in their initializer, which meant that when I use
`add_child(blah, true)` it will calculate a name for it that looks like
"Request", "Request2", "Slot3", etc. That means when I'm debugging and
looking in the remote scene tree I can see all the offers and requests
there and inspect them, which is neat.

The last thing I did was the Slot abstraction which allows me to get rid
of even more state. For the factories I now don't have to keep track of
inventory or ongoing requests, etc. Instead I can just drop 2 requests
into the slot and then subscribe to a signal that tells me when the slot
is empty. That means all requests are fulfilled, where the state of how
many are outstanding is reflected in the children of the slot and they
manage their own state about what they're doing.

So my final flow for the Red Factory is now:
1. Put 2 requests into the slot
2. When slot is empty, start the timer
3. When the timer is done, put an offer into the slot
4. When the offer is gone, put 2 requests into the slot, repeat

So the factory does what I need without having to actually manage that
much. Pretty good!

Ok, so, the last step is Warehouse. It's complicated enough that I'll
probably need to make some changes still, but hopefully it's not too
much trouble...