Refactor Market 3: Warehouse
I didn't have to change as much about the new abstractions as I expected
to, which is nice.

So, Warehouse has 10 slots, and it turns out I actually didn't want the
slot signals I built for the factories here because I actually want to
know what the offers and requests are so I can keep my internal
inventory in sync.

I could rebuild that every time by looking in all the slots and seeing
what offers they have, but I think it's both easier and faster to just
keep the list as we go.

Then I had to actually implement the exclusive_id logic so a single slot
in the warehouse that's able to take both red and blue, say, won't get
both. I need it to work as an either. That allowed my to partially
replace my "taken" array, but I still needed one for the non-exclusive
cases where I just don't want to give the same job to 3 bots.

I considered doing a thing where I use the `object_instance_id` of
anything that didn't have an exclusive ID so they could all go in the
same logic, but the problem there is that I have to subscribe to those
so I can free up the lock later. It just felt a bit excessive to now
have two subscriptions to every single pickup and dropoff in the whole
system just to clear state that didn't need to be cleared, so that gives
me two levels. One globally unique, and the other locally unique.
They're very similar, though.

So all in all, it went pretty well!