Finish Market Migration and Cleanup Market Debug Flags
I had a few bits and bobs left over after the last migration that needed
to be there for compatibility with the old way, but weren't required in
the new way.

One thing I would lose though was the debugging output, which needed
some sort of human-readable string for anything to make sense.

So that moved me to overhaul that crap too, so I now have a set of flags
rather than just one, and I've moved around where these things are
defined so they're more together.

I've also changed the human-readable bits so it's now a full path
meaning I can actually look in the remote tree and find this thing now,
which is way easier to reconcile. Rather than just a list of orders I
now also list the requests and the accepted orders, and I print it nicer
so that it is easier to tell the end of one and the start of the next.

The last thing is something I've wanted to do for a while, but was
necessitated by these extra logs, which is a rate-limit on the market.
This means the bots that are assigned stuff go about their merry way,
but the new assignments only happen every 5 seconds when debugging the
market. This makes it much easier to actually _see_ what's happening and
reason about whether or not that's what I expected, etc, versus a
constant stream of a bunch of empty markets 60 times a second or
whatever.