Construction Costs
Ok, I've now implemented a feature. Building things costs resources, and
then time after that.

For now the cost is hard-coded, but eventually it will likely change
based on what I'm building.
And for now the benefit of building is also hard-coded, but that will
also likely change in the future to be based on some kind of
"construction" intent.

But this is something. I can now setup a whole bunch of intention, and
the little buddies will work on it. Cute!

And then for practical reasons I also added a debug flag that makes the
buildings free so I can test things faster.